Move express middlewares to server dir
